+++ test/runner.go
@@ -1,9 +1,16 @@
 package test
 
+import "sync"
+
 // Runner is the interface responsible for running tests implemented by testing.M.
 type Runner interface {
 	// Run executes the tests and returns with an exit code.
 	Run() int
+}
+
+// runnerFactory creates a new runner.
+type runnerFactory interface {
+	CreateRunner() (Runner, error)
 }
 
 // Runners is a list of test runners.
@@ -40,4 +47,39 @@ func AppendRunner(target Runner, runners ...Runner) Runner {
 	r = append(r, runners...)
 
 	return r
+}
+
+// RunnerRegistry accepts runner factory implementations and creates a runner list from them.
+type RunnerRegistry struct {
+	factories []runnerFactory
+
+	mu sync.Mutex
+}
+
+// Register appends a runner factory to the list.
+// It is safe to call this method from multiple goroutines if necessary.
+func (r *RunnerRegistry) Register(factory runnerFactory) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
+	r.factories = append(r.factories, factory)
+}
+
+// CreateRunner creates test runners from the underlying factories.
+func (r *RunnerRegistry) CreateRunner() (Runner, error) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
+	runners := make(Runners, len(r.factories))
+
+	for index, factory := range r.factories {
+		runner, err := factory.CreateRunner()
+		if err != nil {
+			return nil, err
+		}
+
+		runners[index] = runner
+	}
+
+	return runners, nil
 }