+++ CHANGELOG.md
@@ -11,8 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Runner interface for test runners (already implemented by Godog and Main test runners)
-- Runner list for merging test runners into one
+- Test runner interface for test runners (already implemented by Godog and Main test runners)
+- Test runner list for merging test runners into one
+- Test runner factories
 
 ### Changed